Moved slider to abstract setup screen since the host must specify how many players will be in the game. Maybe fixed memory leak
